+++ iblrig/pydantic_definitions.py
@@ -1,9 +1,25 @@
 from collections import abc
 from datetime import date
 from pathlib import Path
-from typing import Literal
+from typing import Annotated, Literal
 
-from pydantic import AnyUrl, BaseModel, ConfigDict, Field, field_serializer, field_validator
+from pydantic import (
+    AnyUrl,
+    BaseModel,
+    ConfigDict,
+    DirectoryPath,
+    Field,
+    FilePath,
+    PlainSerializer,
+    PositiveFloat,
+    field_serializer,
+    field_validator,
+)
+
+from iblrig.constants import BASE_PATH
+
+FilePath = Annotated[FilePath, PlainSerializer(lambda s: str(s), return_type=str)]
+"""Validate that path exists and is file. Cast to str upon save."""
 
 
 class BunchModel(BaseModel, abc.MutableMapping):
@@ -29,7 +45,7 @@ class BunchModel(BaseModel, abc.MutableMapping):
         return (getattr(self, key) for key in self.keys())
 
     def __delitem__(self, key):
-        raise NotImplementedError()
+        raise NotImplementedError
 
 
 class RigSettings(BunchModel, validate_assignment=True):
@@ -37,14 +53,26 @@ class RigSettings(BunchModel, validate_assignment=True):
     iblrig_local_data_path: Path | None = Field(
         title='IBLRIG local data path', description='The local folder IBLRIG should use for storing data'
     )
+    iblrig_local_subjects_path: DirectoryPath | None = Field(
+        title='IBLRIG full local data path',
+        omit_default=True,
+        default=None,
+        description='An optional full local data folder (including /Subjects)',
+    )
     iblrig_remote_data_path: Path | bool | None = Field(
         title='IBLRIG remote data path', description='The remote folder IBLRIG should use for storing data'
+    )
+    iblrig_remote_subjects_path: Path | None = Field(
+        title='IBLRIG full remote data path',
+        omit_default=True,
+        default=None,
+        description='An optional full remote data folder (including /Subjects)',
     )
     ALYX_USER: str | None = Field(description='Your Alyx username')
     ALYX_URL: AnyUrl | None = Field(title='Alyx URL', description='The URL to your Alyx database')
     ALYX_LAB: str | None = Field(description="Your lab's name as registered on the Alyx database")
 
-    @field_validator('*')
+    @field_validator('ALYX_USER', 'ALYX_LAB')
     def str_must_not_contain_space(cls, v):  # noqa: N805
         if isinstance(v, str) and ' ' in v:
             raise ValueError('must not contain a space')
@@ -88,16 +116,16 @@ class HardwareSettingsScreen(BunchModel):
 class HardwareSettingsSound(BunchModel):
     OUTPUT: Literal['harp', 'xonar', 'hifi', 'sysdefault']
     COM_SOUND: str | None = None
-    # AMP_TYPE: Literal['harp', 'AMP2X15'] | None = None
+    AMP_TYPE: Literal['harp', 'AMP2X15'] | None = None
     # ATTENUATION_DB: float = Field(default=0, le=0)
 
 
 class HardwareSettingsValve(BunchModel):
     WATER_CALIBRATION_DATE: date
-    WATER_CALIBRATION_RANGE: list[float] = Field(min_items=2, max_items=2)  # type: ignore
-    WATER_CALIBRATION_OPEN_TIMES: list[float] = Field(min_items=2)  # type: ignore
-    WATER_CALIBRATION_WEIGHT_PERDROP: list[float] = Field(min_items=2)  # type: ignore
-    FREE_REWARD_VOLUME_UL: float = 1.5
+    WATER_CALIBRATION_RANGE: list[PositiveFloat] = Field(min_items=2, max_items=2)  # type: ignore
+    WATER_CALIBRATION_OPEN_TIMES: list[PositiveFloat] = Field(min_items=2)  # type: ignore
+    WATER_CALIBRATION_WEIGHT_PERDROP: list[float] = Field(PositiveFloat, min_items=2)  # type: ignore
+    FREE_REWARD_VOLUME_UL: PositiveFloat = 1.5
 
 
 class HardwareSettingsScale(BunchModel):
@@ -105,17 +133,52 @@ class HardwareSettingsScale(BunchModel):
 
 
 class HardwareSettingsCamera(BunchModel):
-    BONSAI_WORKFLOW: Path
+    INDEX: int
+    FPS: int | None = Field(
+        title='Camera frame rate',
+        omit_default=True,
+        default=None,
+        description='An optional frame rate (for camera QC only)',
+        ge=0,
+    )
+    WIDTH: int | None = Field(
+        title='Camera frame width',
+        omit_default=True,
+        default=None,
+        description='An optional frame width (for camera QC only)',
+        ge=0,
+    )
+    HEIGHT: int | None = Field(
+        title='Camera frame height',
+        omit_default=True,
+        default=None,
+        description='An optional frame hight (for camera QC only)',
+        ge=0,
+    )
+    SYNC_LABEL: str | None = Field(
+        title='Camera DAQ sync label',
+        omit_default=True,
+        default=None,
+        description='The name of the DAQ channel wired to the camera GPIO',
+    )
 
-    @field_serializer('BONSAI_WORKFLOW')
-    def serialize_path(self, bonsai_workflow: Path, _info):
-        return str(bonsai_workflow)
 
+class HardwareSettingsCameraWorkflow(BunchModel):
+    setup: FilePath | None = Field(
+        title='Optional camera setup workflow',
+        omit_default=True,
+        default=None,
+        description='An optional path to the camera setup Bonsai workflow.',
+    )
+    recording: FilePath = Field(
+        title='Camera recording workflow', description='The path to the Bonsai workflow for camera recording.'
+    )
 
-class HardwareSettingsCameras(BunchModel):
-    left: HardwareSettingsCamera | None
-    right: HardwareSettingsCamera | None = None
-    body: HardwareSettingsCamera | None = None
+    @field_validator('setup', 'recording', mode='before')
+    def valid_path(cls, v):  # noqa: N805
+        if not Path(v).is_absolute():  # assume relative to iblrig repo
+            v = BASE_PATH.joinpath(v)
+        return v
 
 
 class HardwareSettingsMicrophone(BunchModel):
@@ -137,6 +200,6 @@ class HardwareSettings(BunchModel):
     device_sound: HardwareSettingsSound
     device_valve: HardwareSettingsValve
     device_scale: HardwareSettingsScale = HardwareSettingsScale()
-    device_cameras: HardwareSettingsCameras | None = None
+    device_cameras: dict[str, dict[str, HardwareSettingsCameraWorkflow | HardwareSettingsCamera]] | None
     device_microphone: HardwareSettingsMicrophone | None = None
     VERSION: str